kafka_producer/requirements.txt
Faker
kafka-python
syslog-rfc5424-parser